refactor: id 타입 변경

+++ backend/src/main/java/com/ssafy/arttab/artwork/dto/ArtworkSaveRequestDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ArtworkSaveRequestDto { // 작품을 저장하는 요청 Request
 
-    private int writerId;
+    private Long writerId;
     private String title;
     private String desc;
     private String originFileName;
@@ -19,7 +19,7 @@ public class ArtworkSaveRequestDto { // 작품을 저장하는 요청 Request
     private double height;
 
     @Builder
-    public ArtworkSaveRequestDto (int writerId, String title, String desc, String originFileName, String saveFileName,
+    public ArtworkSaveRequestDto (Long writerId, String title, String desc, String originFileName, String saveFileName,
                                   String saveFolder, int size, double width, double height){
         this.writerId=writerId;
         this.title=title;
